requirements.txt
numpy==1.21.6
pandas==1.3.5
opencv-python==4.7.0.72
matplotlib==3.1.0
tensorflow==2.11.0
scikit-learn==1.0.2
streamlit==1.20.0
Pillow==9.4.0
face-recognition==1.3.0